extract MultiSelectComboBoxI18n into separate class
