+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/main/java/com/vaadin/flow/component/combobox/MultiSelectComboBox.java
@@ -35,7 +35,6 @@ import elemental.json.JsonArray;
 import elemental.json.JsonObject;
 import elemental.json.JsonType;
 
-import java.io.Serializable;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
@@ -433,134 +432,6 @@ public class MultiSelectComboBox<TItem>
             if (jsonObject.get(key).getType() == JsonType.NULL) {
                 jsonObject.remove(key);
             }
-        }
-    }
-
-    /**
-     * Class for localization of the {@link MultiSelectComboBox}
-     */
-    public static class MultiSelectComboBoxI18n implements Serializable {
-        private String cleared;
-        private String focused;
-        private String selected;
-        private String deselected;
-        private String total;
-
-        /**
-         * The text that is announced by screen readers when the clear button is
-         * clicked.
-         * <p>
-         * The value is {@code null} by default, which means the default value
-         * of the web component will be used.
-         */
-        public String getCleared() {
-            return cleared;
-        }
-
-        /**
-         * Sets the text that is announced by screen readers when the clear
-         * button is clicked.
-         *
-         * @return this instance for method chaining
-         */
-        public MultiSelectComboBoxI18n setCleared(String cleared) {
-            this.cleared = cleared;
-            return this;
-        }
-
-        /**
-         * The text that is announced by screen readers when a chip is focused.
-         * <p>
-         * The value is {@code null} by default, which means the default value
-         * of the web component will be used.
-         */
-        public String getFocused() {
-            return focused;
-        }
-
-        /**
-         * Sets the text that is announced by screen readers when a chip is
-         * focused. The label of the chip will be prepended to this text.
-         *
-         * @return this instance for method chaining
-         */
-        public MultiSelectComboBoxI18n setFocused(String focused) {
-            this.focused = focused;
-            return this;
-        }
-
-        /**
-         * The text that is announced by screen readers when an item is added to
-         * the selection.
-         * <p>
-         * The value is {@code null} by default, which means the default value
-         * of the web component will be used.
-         */
-        public String getSelected() {
-            return selected;
-        }
-
-        /**
-         * Sets the text that is announced by screen readers when an item is
-         * added to the selection. The label of the item will be prepended to
-         * this text.
-         *
-         * @return this instance for method chaining
-         */
-        public MultiSelectComboBoxI18n setSelected(String selected) {
-            this.selected = selected;
-            return this;
-        }
-
-        /**
-         * The text that is announced by screen readers when an item is removed
-         * from the selection.
-         * <p>
-         * The value is {@code null} by default, which means the default value
-         * of the web component will be used.
-         */
-        public String getDeselected() {
-            return deselected;
-        }
-
-        /**
-         * Sets the text that is announced by screen readers when an item is
-         * removed from the selection. The label of the item will be prepended
-         * to this text.
-         *
-         * @return this instance for method chaining
-         */
-        public MultiSelectComboBoxI18n setDeselected(String deselected) {
-            this.deselected = deselected;
-            return this;
-        }
-
-        /**
-         * The text that is announced by screen readers to inform about the
-         * total number of selected items.
-         * <p>
-         * The value is {@code null} by default, which means the default value
-         * of the web component will be used.
-         */
-        public String getTotal() {
-            return total;
-        }
-
-        /**
-         * Sets the text that is announced by screen readers to inform about the
-         * total number of selected items. The string must contain a `{count}`
-         * placeholder that will be replaced with the actual count of selected
-         * items by the component.
-         *
-         * @return this instance for method chaining
-         */
-        public MultiSelectComboBoxI18n setTotal(String total) {
-            if (total != null && !total.contains("{count}")) {
-                throw new IllegalArgumentException(
-                        "Text must contain a {count} placeholder");
-            }
-            this.total = total;
-            return this;
         }
     }
 }
