+++ vaadin-combo-box-flow-parent/vaadin-combo-box-flow/src/test/java/com/vaadin/flow/component/combobox/MultiSelectComboBoxI18nTest.java
@@ -15,7 +15,7 @@ public class MultiSelectComboBoxI18nTest {
 
     @Test
     public void setI18n() {
-        MultiSelectComboBox.MultiSelectComboBoxI18n i18n = new MultiSelectComboBox.MultiSelectComboBoxI18n()
+        MultiSelectComboBoxI18n i18n = new MultiSelectComboBoxI18n()
                 .setCleared("All entries removed");
         comboBox.setI18n(i18n);
 
@@ -29,13 +29,13 @@ public class MultiSelectComboBoxI18nTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void setTotalWithoutCountPlaceholder_throws() {
-        MultiSelectComboBox.MultiSelectComboBoxI18n i18n = new MultiSelectComboBox.MultiSelectComboBoxI18n();
+        MultiSelectComboBoxI18n i18n = new MultiSelectComboBoxI18n();
         i18n.setTotal("entries selected");
     }
 
     @Test()
     public void setTotalWithCountPlaceholder_doesNotThrow() {
-        MultiSelectComboBox.MultiSelectComboBoxI18n i18n = new MultiSelectComboBox.MultiSelectComboBoxI18n();
+        MultiSelectComboBoxI18n i18n = new MultiSelectComboBoxI18n();
         i18n.setTotal("{count} entries selected");
 
         Assert.assertEquals("{count} entries selected", i18n.getTotal());
